ID-299 fixing tests etc#minor
