--- service-front/module/Application/src/Controller/HealthCheckController.php
@@ -1,62 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Application\Controller;
-
-use Application\Contracts\OpgApiServiceInterface;
-use Application\Services\SiriusApiService;
-use Application\Views\JsonModel;
-use Laminas\Mvc\Controller\AbstractActionController;
-
-/**
- * @psalm-suppress PropertyNotSetInConstructor
- * @psalm-suppress InvalidArgument
- * @psalm-suppress UnusedProperty
- * Needed here due to false positive from Laminas’s uninitialised properties
- */
-class HealthCheckController extends AbstractActionController
-{
-    public function __construct(
-        private readonly OpgApiServiceInterface $opgApiService,
-        private readonly SiriusApiService $siriusApiService,
-    )
-    {
-    }
-
-    public function healthCheckAction(): JsonModel
-    {
-        return new JsonModel([
-            'OK' => true
-        ]);
-    }
-
-    public function healthCheckServiceAction(): JsonModel
-    {
-        $ok = true;
-
-        $siriusResponse = $this->siriusApiService->checkAuth($this->getRequest());
-
-        if ($siriusResponse !== true) {
-            $ok = false;
-        }
-
-        $apiResponse = $this->opgApiService->healthCheck();
-
-        if ($apiResponse !== true) {
-            $ok = false;
-        }
-
-        return new JsonModel([
-            'OK' => $ok,
-            'dependencies' => [
-                'sirius' => [
-                    'ok' => $siriusResponse
-                ],
-                'api' => [
-                    'ok' => $apiResponse
-                ]
-            ]
-        ]);
-    }
-}
