+++ service-front/module/Application/test/Controller/IndexControllerTest.php
@@ -222,4 +222,32 @@ class IndexControllerTest extends AbstractHttpControllerTestCase
             $this->getResponse()->getBody()
         );
     }
+
+    public function testHealthCheckAction(): void
+    {
+        $this->dispatch('/health-check', 'GET');
+        $this->assertResponseStatusCode(200);
+    }
+
+    public function testHealthCheckServiceAction(): void
+    {
+        $siriusApiService = $this->createMock(SiriusApiService::class);
+        $opgApiService = $this->createMock(OpgApiService::class);
+
+        $serviceManager = $this->getApplicationServiceLocator();
+        $serviceManager->setService(SiriusApiService::class, $siriusApiService);
+        $serviceManager->setService(OpgApiService::class, $opgApiService);
+
+        $siriusApiService->expects($this->once())
+            ->method('checkAuth')
+            ->willReturn(true);
+
+        $opgApiService->expects($this->once())
+            ->method('healthCheck')
+
+            ->willReturn(true);
+
+        $this->dispatch('/health-check-service', 'GET');
+        $this->assertResponseStatusCode(200);
+    }
 }
