+++ service-front/module/Application/view/application/pages/healthcheck/healthcheck.twig
@@ -1,0 +1,1 @@
+{{ status }}
